changed code layout, test works
