+++ src/autonomous/test_auton.py
@@ -10,8 +10,8 @@ class AutonDrive(wpilib.command.Command):
         print("Initializing.")
 
     def execute(self):
-        Robot.dt.driveTWH(0.5, 0.0, 0.0)
+        Robot.dt.drive_TWH(0.5, 0.0, 0.0)
         if Robot.encoder.get() > 10:
-            Robot.dt.driveTWH(0.0, 0.0, 0.0)
+            Robot.dt.drive_TWH(0.0, 0.0, 0.0)
     def isFinished(self):
         return Robot.encoder.get() > 10
